Refactor: decouple scheduler from websitemonitor
The frequency_online and frequency_offline options are cute, maybe
slightly useful, but ultimately they dramatically change the
architecture by coupling the scheduler with the website monitor,
primarily because the result of the check determines the scheduling.

This complicates or prevents a number of things, such as adding a
queue between the scheduling and the execution (which would help
scale), and the benefit is too small to justify the complication.

This change also decouples the scheduler implementation from the system,
adding a facade and removing all scheduler interaction from the
WebsiteMonitor class.